# target/classes/swagger.json
{
	"swagger": "2.0",
	"info": 
	{
		"description": "This is a sample server Petstore server.  You can find out more about Swagger at [http://swagger.io](http://swagger.io) or on [irc.freenode.net, #swagger](http://swagger.io/irc/).  For this sample, you can use the api key `special-key` to test the authorization filters.",
		"version": "1.0.0",
		"title": "Swagger Petstore",
		"termsOfService": "http://swagger.io/terms/",
		"contact": 
		{
			"email": "apiteam@swagger.io"
		},

		"license": 
		{
			"name": "Apache 2.0",
			"url": "http://www.apache.org/licenses/LICENSE-2.0.html"
		}
	},
	"host": "petstore.swagger.io",
	"basePath": "/v2",
...
	"schemes": 
	[
		"https",
		"http"
	],
	"paths": 
...
		"/store/inventory": 
		{
			"get": 
			{
				"tags": 
				[
					"store"
				],

				"summary": "Returns pet inventories by status",
				"description": "Returns a map of status codes to quantities",
				"operationId": "getInventory",
				"produces": 
				[
					"application/json"
				],

				"parameters": 
				[
					
				],

				"responses": 
				{
					"200": 
					{
						"description": "successful operation",
						"schema": 
						{
							"type": "object",
							"additionalProperties": 
							{
								"type": "integer",
								"format": "int32"
							}
						}
					}
				},

				"security": 
				[
					{
						"api_key": 
						[
							
						]
					}
				]
			}
		},
		"/store/order": 
		{
			"post": 
			{
				"tags": 
				[
					"store"
				],

				"summary": "Place an order for a pet",
				"description": "",
				"operationId": "placeOrder",
				"produces": 
				[
					"application/xml",
					"application/json"
				],

				"parameters": 
				[
					{
						"in": "body",
						"name": "body",
						"description": "order placed for purchasing the pet",
						"required": true,
						"schema": 
						{
							"$ref": "#/definitions/Order"
						}
					}
				],

				"responses": 
				{
					"200": 
					{
						"description": "successful operation",
						"schema": 
						{
							"$ref": "#/definitions/Order"
						}
					},

					"400": 
					{
						"description": "Invalid Order"
					}
				}
			}
		},
		"/store/order/{orderId}": 
		{
			"get": 
			{
				"tags": 
				[
					"store"
				],

				"summary": "Find purchase order by ID",
				"description": "For valid response try integer IDs with value >= 1 and <= 10. Other values will generated exceptions",
				"operationId": "getOrderById",
				"produces": 
				[
					"application/xml",
					"application/json"
				],

				"parameters": 
				[
					{
						"name": "orderId",
						"in": "path",
						"description": "ID of pet that needs to be fetched",
						"required": true,
						"type": "integer",
						"maximum": 10.0,
						"minimum": 1.0,
						"format": "int64"
					}
				],

				"responses": 
				{
					"200": 
					{
						"description": "successful operation",
						"schema": 
						{
							"$ref": "#/definitions/Order"
						}
					},

					"400": 
					{
						"description": "Invalid ID supplied"
					},

					"404": 
					{
						"description": "Order not found"
					}
				}
			},

			"delete": 
			{
				"tags": 
				[
					"store"
				],

				"summary": "Delete purchase order by ID",
				"description": "For valid response try integer IDs with positive integer value. Negative or non-integer values will generate API errors",
				"operationId": "deleteOrder",
				"produces": 
				[
					"application/xml",
					"application/json"
				],

				"parameters": 
				[
					{
						"name": "orderId",
						"in": "path",
						"description": "ID of the order that needs to be deleted",
						"required": true,
						"type": "integer",
						"minimum": 1.0,
						"format": "int64"
					}
				],

				"responses": 
				{
					"400": 
					{
						"description": "Invalid ID supplied"
					},

					"404": 
					{
						"description": "Order not found"
					}
				}
			}
		},
		"/user": 
		{
			"post": 
			{
				"tags": 
				[
					"user"
				],

				"summary": "Create user",
				"description": "This can only be done by the logged in user.",
				"operationId": "createUser",
				"produces": 
				[
					"application/xml",
					"application/json"
				],

				"parameters": 
				[
					{
						"in": "body",
						"name": "body",
						"description": "Created user object",
						"required": true,
						"schema": 
						{
							"$ref": "#/definitions/User"
						}
					}
				],

				"responses": 
				{
					"default": 
					{
						"description": "successful operation"
					}
				}
			}
		},
		"/user/createWithArray": 
		{
			"post": 
			{
				"tags": 
				[
					"user"
				],

				"summary": "Creates list of users with given input array",
				"description": "",
				"operationId": "createUsersWithArrayInput",
				"produces": 
				[
					"application/xml",
					"application/json"
				],

				"parameters": 
				[
					{
						"in": "body",
						"name": "body",
						"description": "List of user object",
						"required": true,
						"schema": 
						{
							"type": "array",
							"items": 
							{
								"$ref": "#/definitions/User"
							}
						}
					}
				],

				"responses": 
				{
					"default": 
					{
						"description": "successful operation"
					}
				}
			}
		},
		"/user/createWithList": 
		{
			"post": 
			{
				"tags": 
				[
					"user"
				],

				"summary": "Creates list of users with given input array",
				"description": "",
				"operationId": "createUsersWithListInput",
				"produces": 
				[
					"application/xml",
					"application/json"
				],

				"parameters": 
				[
					{
						"in": "body",
						"name": "body",
						"description": "List of user object",
						"required": true,
						"schema": 
						{
							"type": "array",
							"items": 
							{
								"$ref": "#/definitions/User"
							}
						}
					}
				],

				"responses": 
				{
					"default": 
					{
						"description": "successful operation"
					}
				}
			}
		},
		"/user/login": 
		{
			"get": 
			{
				"tags": 
				[
					"user"
				],

				"summary": "Logs user into the system",
				"description": "",
				"operationId": "loginUser",
				"produces": 
				[
					"application/xml",
					"application/json"
				],

				"parameters": 
				[
					{
						"name": "username",
						"in": "query",
						"description": "The user name for login",
						"required": true,
						"type": "string"
					},

					{
						"name": "password",
						"in": "query",
						"description": "The password for login in clear text",
						"required": true,
						"type": "string"
					}
				],

				"responses": 
				{
					"200": 
					{
						"description": "successful operation",
						"schema": 
						{
							"type": "string"
						},

						"headers": 
						{
							"X-Rate-Limit": 
							{
								"type": "integer",
								"format": "int32",
								"description": "calls per hour allowed by the user"
							},

							"X-Expires-After": 
							{
								"type": "string",
								"format": "date-time",
								"description": "date in UTC when token expires"
							}
						}
					},

					"400": 
					{
						"description": "Invalid username/password supplied"
					}
				}
			}
		},
...
		"/user/{username}": 
		{
			"get": 
			{
				"tags": 
				[
					"user"
				],

				"summary": "Get user by user name",
				"description": "",
				"operationId": "getUserByName",
				"produces": 
				[
					"application/xml",
					"application/json"
				],

				"parameters": 
				[
					{
						"name": "username",
						"in": "path",
						"description": "The name that needs to be fetched. Use user1 for testing. ",
						"required": true,
						"type": "string"
					}
				],

				"responses": 
				{
					"200": 
					{
						"description": "successful operation",
						"schema": 
						{
							"$ref": "#/definitions/User"
						}
					},

					"400": 
					{
						"description": "Invalid username supplied"
					},

					"404": 
					{
						"description": "User not found"
					}
				}
			},

			"put": 
			{
				"tags": 
				[
					"user"
				],

				"summary": "Updated user",
				"description": "This can only be done by the logged in user.",
				"operationId": "updateUser",
				"produces": 
				[
					"application/xml",
					"application/json"
				],

				"parameters": 
				[
					{
						"name": "username",
						"in": "path",
						"description": "name that need to be updated",
						"required": true,
						"type": "string"
					},

					{
						"in": "body",
						"name": "body",
						"description": "Updated user object",
						"required": true,
						"schema": 
						{
							"$ref": "#/definitions/User"
						}
					}
				],

				"responses": 
				{
					"400": 
					{
						"description": "Invalid user supplied"
					},

					"404": 
					{
						"description": "User not found"
					}
				}
			},

			"delete": 
			{
				"tags": 
				[
					"user"
				],

				"summary": "Delete user",
				"description": "This can only be done by the logged in user.",
				"operationId": "deleteUser",
				"produces": 
				[
					"application/xml",
					"application/json"
				],

				"parameters": 
				[
					{
						"name": "username",
						"in": "path",
						"description": "The name that needs to be deleted",
						"required": true,
						"type": "string"
					}
				],

				"responses": 
				{
					"400": 
					{
						"description": "Invalid username supplied"
					},

					"404": 
					{
						"description": "User not found"
					}
				}
			}
		}
	},
	"securityDefinitions": 
	{
		"petstore_auth": 
		{
			"type": "oauth2",
			"authorizationUrl": "https://petstore.swagger.io/oauth/authorize",
			"flow": "implicit",
			"scopes": 
			{
				"write:pets": "modify pets in your account",
				"read:pets": "read your pets"
			}
		},

		"api_key": 
		{
			"type": "apiKey",
			"name": "api_key",
			"in": "header"
		}
	},
	"definitions": 
...
		"Order": 
		{
			"type": "object",
			"properties": 
			{
				"id": 
				{
					"type": "integer",
					"format": "int64"
				},

				"petId": 
				{
					"type": "integer",
					"format": "int64"
				},

				"quantity": 
				{
					"type": "integer",
					"format": "int32"
				},

				"shipDate": 
				{
					"type": "string",
					"format": "date-time"
				},

				"status": 
				{
					"type": "string",
					"description": "Order Status",
					"enum": 
					[
						"placed",
						"approved",
						"delivered"
					]
				},

				"complete": 
				{
					"type": "boolean",
					"default": false
				}
			},

			"xml": 
			{
				"name": "Order"
			}
		},
		"User": 
		{
			"type": "object",
			"properties": 
			{
				"id": 
				{
					"type": "integer",
					"format": "int64"
				},

				"username": 
				{
					"type": "string"
				},

				"firstName": 
				{
					"type": "string"
				},

				"lastName": 
				{
					"type": "string"
				},

				"email": 
				{
					"type": "string"
				},

				"password": 
				{
					"type": "string"
				},

				"phone": 
				{
					"type": "string"
				},

				"userStatus": 
				{
					"type": "integer",
					"format": "int32",
					"description": "User Status"
				}
			},

			"xml": 
			{
				"name": "User"
			}
		},
...
	},
...
}